Fixed device restart caused by starting Activity under split-screen root task.
The root task used for split screen contains two or more sub-tasks. In this case, the activity should not be allowed to be started under the root task, which will cause a null pointer and cause the device to restart. Therefore, when the activity is started by specifying the taskid, check that the task is leaftask.

Google: 3079172
Change-Id: I095feb549ce2d3a4e61d999a342f64bea8264288